Add reference material note
I didn't want to potentially point to this as a portfolio project because I'm using the reference repository in its entirety.
These commits do highlight the steps required to add deployment on fly.io with CI/CD through GitHub Actions but those changes are largely following the given commands.

This is a good reference for understanding relationships in Ecto and we spent class time working through each iteration. In the future, it would be useful to add some embellishments or pieces of flare as it were. I did this with the TodoMVC project so it makes sense to do it here when we have more time. Our capstone project will be entirely novel though.